# packages/api/cfn/swagger.yml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Clean Arch Todos API
...
paths:
...
  /lists:
    post:
      tags:
        - todos
      summary: Create List
      operationId: createList
      parameters:
        - name: X-Request-Id
          in: header
          required: true
          description: Request identifier
          schema:
            type: string
      requestBody:
        description: List parameters
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListParams"
      responses:
        201:
          description: Create List Success
      x-amazon-apigateway-integration:
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${CreateList.Arn}/invocations
        passthroughBehavior: when_no_match
        httpMethod: POST
        type: aws_proxy
components:
  schemas:
...
    CreateListParams:
      type: object
      properties:
        name:
          type: string
        allowExpired:
          type: boolean
...